Update dni_custom.js
There is no "call" property on an href. I am changing this to check that the screen size is smaller than mobile before the console message, and to replace the inner html if the screen is laptop size or bigger. This should also catch any tablets that use the desktop version when viewed from a landscape perspective.